# speed up api calls for transaction status messages