--- a/etc/module.xml
+++ b/etc/module.xml
@@ -25,7 +25,7 @@
  */
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="../../../../../lib/internal/Magento/Framework/Module/etc/module.xsd">
-    <module name="Payone_Core" setup_version="2.3.1">
+    <module name="Payone_Core" setup_version="2.3.2">
         <sequence>
             <module name="Magento_Quote" />
             <module name="Magento_Sales" />