- Transaction status messages identify the according order by the txid. The column had no index
and therefore those requests may have consumed some time when a lot of transactions are held in
this table.
- Remove wrong duplicated comment

In our scenario the order lookup in transaction status processing could be reduced from 1.4 to 1.7 seconds (no query_cache) to 200ms to 300ms.

*Repro*
1. Run `bin/magento setup:upgrade`
1. check structure of table `payone_protocol_api`
   1. An index `PAYONE_PROTOCOL_API_TXID` exists