@steros 2018-07-27
This might as well affect the table `payone_protocol_transactionstatus`. I see that this table is queried by customerid and txid as well. Both columns do not have an index.