@fjbender 2018-07-27
Thanks @steros, added indices to that table as well with b00cbb4.