@fjbender 2018-07-27
@hreinberger Please review